ChessBot/requirements.txt
discord.py
chess 
fentoboardimage